third_party/WebKit/LayoutTests/fast/ime/candidatewindowevent-expected.txt

This tests if candidatewindow events work for InputMethodContext.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.type is "candidatewindowshow"
PASS event.type is "candidatewindowupdate"
PASS event.type is "candidatewindowhide"
PASS event.type is "candidatewindowshow"
PASS event.type is "candidatewindowupdate"
PASS event.type is "candidatewindowhide"
PASS successfullyParsed is true

TEST COMPLETE
